requirements.txt
absl-py==2.0.0
appnope==0.1.3
asttokens==2.4.1
astunparse==1.6.3
audioread==3.0.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.46.0
fsspec==2023.12.2
gast==0.5.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.10.0
huggingface-hub==0.19.4
idna==3.6
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.1
llvmlite==0.41.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.7
nest-asyncio==1.5.8
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
parso==0.8.3
patsy==0.5.4
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
pooch==1.8.0
praat-parselmouth==0.4.3
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
stack-data==0.6.3
statsmodels==0.14.1
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-hub==0.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformers==4.36.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
wrapt==1.14.1
